Code/requirements.txt
flask
flask_restful
flask_pymongo
python-crontab
dnspython
requests
celery-redbeat
celery[redis]
flower
python-nmap
python-whois
ipwhois